+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added 'remote job' detection
+- Added 'clear-cache' option to the CLI
 
 ### Changed